remove spatial_full_distortion_kernel_func_ptr_array

+++ Source/Lib/Common/ASM_AVX2/EbPictureOperators_AVX2.h
@@ -124,7 +124,7 @@ extern "C" {
         return Hadd32_SSE2_INTRIN(sum);
     }
 
-    uint64_t spatial_full_distortion_kernel4x_n_avx2_intrin(
+    uint32_t spatial_full_distortion_kernel4x_n_avx2_intrin(
         uint8_t   *input,
         uint32_t   input_stride,
         uint8_t   *recon,
@@ -132,7 +132,7 @@ extern "C" {
         uint32_t   area_width,
         uint32_t   area_height);
 
-    uint64_t spatial_full_distortion_kernel8x_n_avx2_intrin(
+    uint32_t spatial_full_distortion_kernel8x_n_avx2_intrin(
         uint8_t   *input,
         uint32_t   input_stride,
         uint8_t   *recon,
@@ -140,7 +140,7 @@ extern "C" {
         uint32_t   area_width,
         uint32_t   area_height);
 
-    uint64_t spatial_full_distortion_kernel16x_n_avx2_intrin(
+    uint32_t spatial_full_distortion_kernel16x_n_avx2_intrin(
         uint8_t   *input,
         uint32_t   input_stride,
         uint8_t   *recon,
@@ -148,7 +148,7 @@ extern "C" {
         uint32_t   area_width,
         uint32_t   area_height);
 
-    uint64_t spatial_full_distortion_kernel32x_n_avx2_intrin(
+    uint32_t spatial_full_distortion_kernel32x_n_avx2_intrin(
         uint8_t   *input,
         uint32_t   input_stride,
         uint8_t   *recon,
@@ -156,7 +156,7 @@ extern "C" {
         uint32_t   area_width,
         uint32_t   area_height);
 
-    uint64_t spatial_full_distortion_kernel64x_n_avx2_intrin(
+    uint32_t spatial_full_distortion_kernel64x_n_avx2_intrin(
         uint8_t   *input,
         uint32_t   input_stride,
         uint8_t   *recon,
@@ -164,7 +164,7 @@ extern "C" {
         uint32_t   area_width,
         uint32_t   area_height);
 
-    uint64_t spatial_full_distortion_kernel128x_n_avx2_intrin(
+    uint32_t spatial_full_distortion_kernel128x_n_avx2_intrin(
         uint8_t   *input,
         uint32_t   input_stride,
         uint8_t   *recon,
