+++ Source/Lib/Common/ASM_SSE2/EbPictureOperators_SSE2.h
@@ -164,7 +164,7 @@ extern "C" {
         return _mm_cvtsi128_si32(dst1);
     }
 
-    uint64_t spatial_full_distortion_kernel4x_n_sse2_intrin(
+    uint32_t spatial_full_distortion_kernel4x_n_sse2_intrin(
         uint8_t   *input,
         uint32_t   input_stride,
         uint8_t   *recon,
@@ -172,7 +172,7 @@ extern "C" {
         uint32_t   area_width,
         uint32_t   area_height);
 
-    uint64_t spatial_full_distortion_kernel8x_n_sse2_intrin(
+    uint32_t spatial_full_distortion_kernel8x_n_sse2_intrin(
         uint8_t   *input,
         uint32_t   input_stride,
         uint8_t   *recon,
@@ -180,7 +180,7 @@ extern "C" {
         uint32_t   area_width,
         uint32_t   area_height);
 
-    uint64_t spatial_full_distortion_kernel16x_n_sse2_intrin(
+    uint32_t spatial_full_distortion_kernel16x_n_sse2_intrin(
         uint8_t   *input,
         uint32_t   input_stride,
         uint8_t   *recon,
@@ -188,7 +188,7 @@ extern "C" {
         uint32_t   area_width,
         uint32_t   area_height);
 
-    uint64_t spatial_full_distortion_kernel32x_n_sse2_intrin(
+    uint32_t spatial_full_distortion_kernel32x_n_sse2_intrin(
         uint8_t   *input,
         uint32_t   input_stride,
         uint8_t   *recon,
@@ -196,7 +196,7 @@ extern "C" {
         uint32_t   area_width,
         uint32_t   area_height);
 
-    uint64_t spatial_full_distortion_kernel64x_n_sse2_intrin(
+    uint32_t spatial_full_distortion_kernel64x_n_sse2_intrin(
         uint8_t   *input,
         uint32_t   input_stride,
         uint8_t   *recon,
@@ -204,7 +204,7 @@ extern "C" {
         uint32_t   area_width,
         uint32_t   area_height);
 
-    uint64_t spatial_full_distortion_kernel128x_n_sse2_intrin(
+    uint32_t spatial_full_distortion_kernel128x_n_sse2_intrin(
         uint8_t   *input,
         uint32_t   input_stride,
         uint8_t   *recon,
